+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Unreleased
 
-* Updated OpenTelemetry core component version(s) to `1.15.2`.
-  ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
-
 * Limit how much of the response body is read when export fails using the HTTP
   JSON transport and informational logging is enabled.
   ([#4117](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4117))
+
+* Updated OpenTelemetry core component version(s) to `1.15.3`.
+  ([#4166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4166))
 
 ## 1.15.0