+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.15.1
+
+Released 2026-Apr-21
+
 * Limit how much of the response body is read when export fails using the HTTP
   JSON transport and informational logging is enabled.
   ([#4117](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4117))